bridge: Re-arrange the physical and virtual memory maps
In particular, this causes us to actually have the kernel memory paged
into the right place in the guest so it can start executing code.

However, there seems to still be a problem when running the kernel where
shortly after an IO operation the VCPU triple-faults. This remains to be
explained, and so for now the output parts of the "kernel" stub are
disabled and it just immediately halts, allowing us to run to completion.